pyproject.toml
[project]
name = "arbiter"
version = "0.1.0"
description = "Ingestion → Signal → Execution Framework"
requires-python = ">=3.10"
dependencies = [
    "discord.py>=2.4.0",
    "requests>=2.28.0",
    "python-dotenv>=1.0.0",
    "yfinance>=0.2.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "telethon>=1.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
